+++ components/content/Price.vue
@@ -1,7 +1,8 @@
 <template>
-  <span>{{ formatPrice(v) }}</span>
+  <span>{{ formatPrice(v, locale) }}</span>
 </template>
 
 <script setup lang="ts">
 defineProps<{ v: string | number }>()
+const { locale } = useI18n()
 </script>